Refactor internal matching routines
Previously, we had a macro for defining functions to check whether any of
the keywords from a given trie were found at a particular starting point in
the target string; ->match called that function repeatedly in a loop.

We now have separate functions for the three variants (with ->match calling
the appropriate function just once); simpler macrology is used to factor out
the common "advance to the next target position and trie node, in lock-step"
code.

While the goal for this change is to ease the implementation for a larger
forthcoming change, it actually makes ->match about 1% faster (presumably
because there's less function-call overhead).  Win!