New algorithm: Aho-Corasick matching
An Aho-Corasick matcher is a trie with one extra feature: each node
additionally has a "fail" pointer.  When matching, if there is no onward
edge from the current node for the current character, we follow the fail
pointer to a different node and try again; the new node is the deepest other
node in the trie that corresponds to the characters we've already
considered.

For example, suppose we're looking for keys "sheep" and "heather" in the
string "sheather".  We initially start down the trie following the s-h-e of
"sheather", but then there's no onward edge labelled "a".  The fail pointer
goes to the node that you'd get to by following h-e from the root.  Then we
look again for an onward edge labelled "a"; this time, we find it, and carry
on until we've found a complete match.

The advantage of this algorithm is that each character of the target string
is considered only once (whereas, in pathological cases, the naive trie
matching algorithm will consider some characters of the string as many times
as the length of the longest key being searched for).  The disadvantage,
though, is that we need an extra pointer in each node; it turns out that,
for the patterns used in the test suite, the increased memory usage actually
reduces performance on a 64-bit machine.  However, I have a cunning plan for
fixing that.