Reduce maximum combined size of nodes
This allows us to use a U32 instead of a size_t, on systems where the latter
is 64-bit; the memory saving from that is significant.  The disadvantage, of
course, is that you can no longer use TMFA for giant search problems.  But
you still get 4 GiB of node data, which in practice is several hundred megs
of key data; that's quite a lot.