Shrink node structure by one byte
Instead of storing the size of the "next" array, including the initial
statically allocated member, we store the size of the dynamically allocated
portion of the array.  This means that, even if a node contains edges
labelled both '\0' and '\xFF', two bytes will suffice to store the min and
size (rather than three).

This doesn't yet shrink the node structure, because there's an additional
byte used for the "final" flag.  But that's going to be removed in the next
commit.