Use less bit-munging in 32-bit mode
Alignment considerations mean that, when a pseudo-pointer is 32 bits long,
we always have 32 bits of space in each node header.  In that situation,
therefore, there's nothing to be gained from encoding the "final" flag in
the same struct member as the fail pointer.  So use the more obvious
implementation in 32-bit, with macros to determine what to do.