Return to storing array size, not allocated size
My reason for storing the size of just the allocated part of the
variable-length array in each `struct node` was to guarantee that the
maximum possible value would fit in an unsigned 8-bit value, on the
assumption that it could go up to 256.  However, that assumption is false:
all supra-ASCII data is converted to UTF-8, and the byte 0xFF can never
appear in well-formed UTF-8.  (In fact, the largest value you can get for
standard Unicode characters, including those that haven't yet been
allocated, is 0xF4, which appears as the leading byte of U+100000 and
greater.

So return to storing the total size; this removes a small amount of
arithmetic in the innermost matching loop, so it's valuable.  And add tests
for (a) malformed UTF-8, since that's the only thing that could break this;
and (b) apparently-pathological sets of keys that include both 0x00 and
0xFF.