Massively reduce memory usage for trie creation
Previously, we had an entire separate trie implementation used only during
construction.  It was essentially the same as the non-Aho-Corasick trie, but
with each node containing a static array of 256 machine pointers, one for
every possible edge; so each node was more than 1 KiB on 32-bit machines (or
2 KiB on 64-bit machines).

The extra "big" trie was effectively used only for calculating the minimum
and maximum edge labels at a given node.  So instead we now use a hash that
maps from keyword prefixes to an integer that encodes the minimum and
maximum directly.  This reduces temporary memory usage during construction
by a factor of about 20.